requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
absl-py=2.1.0=pypi_0
aiohttp=3.9.3=pypi_0
aiosignal=1.3.1=pypi_0
antlr4-python3-runtime=4.8=pypi_0
argcomplete=3.2.3=pypi_0
argoverse=1.1.0=pypi_0
async-timeout=4.0.3=pypi_0
attrs=23.2.0=pypi_0
av=12.0.0=pypi_0
av2=0.2.1=pypi_0
bzip2=1.0.8=hd590300_5
ca-certificates=2024.2.2=hbcca054_0
certifi=2024.2.2=pypi_0
click=8.1.7=pypi_0
colorlog=6.8.2=pypi_0
colour=0.1.5=pypi_0
commonmark=0.9.1=pypi_0
cycler=0.12.1=pypi_0
descartes=1.1.0=pypi_0
docker-pycreds=0.4.0=pypi_0
fire=0.6.0=pypi_0
frozenlist=1.4.1=pypi_0
fsspec=2024.3.1=pypi_0
future=1.0.0=pypi_0
gitdb=4.0.11=pypi_0
gitpython=3.1.43=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
h5py=3.10.0=pypi_0
huggingface-hub=0.22.2=pypi_0
hydra-core=1.1.0=pypi_0
idna=3.6=pypi_0
jinja2=3.1.3=pypi_0
joblib=1.3.2=pypi_0
jsonschema=4.21.1=pypi_0
jsonschema-specifications=2023.12.1=pypi_0
kiwisolver=1.4.5=pypi_0
lapsolver=1.1.0=pypi_0
ld_impl_linux-64=2.40=h41732ed_0
libffi=3.4.2=h7f98852_5
libgcc-ng=13.2.0=h807b86a_5
libgomp=13.2.0=h807b86a_5
libnsl=2.0.1=hd590300_0
libsqlite=3.45.2=h2797004_0
libuuid=2.38.1=h0b41bf4_0
libxcrypt=4.4.36=hd590300_1
libzlib=1.2.13=hd590300_5
lightning-utilities=0.3.0=pypi_0
llvmlite=0.39.1=pypi_0
matplotlib=3.4.3=pypi_0
motmetrics=1.1.3=pypi_0
mpmath=1.3.0=pypi_0
multidict=6.0.5=pypi_0
natten=0.14.2+torch111cu113=pypi_0
ncurses=6.4.20240210=h59595ed_0
networkx=3.1=pypi_0
nox=2024.3.2=pypi_0
numba=0.56.4=pypi_0
numpy=1.21.1=pypi_0
nvidia-cublas-cu12=12.1.3.1=pypi_0
nvidia-cuda-cupti-cu12=12.1.105=pypi_0
nvidia-cuda-nvrtc-cu12=12.1.105=pypi_0
nvidia-cuda-runtime-cu12=12.1.105=pypi_0
nvidia-cudnn-cu12=8.9.2.26=pypi_0
nvidia-cufft-cu12=11.0.2.54=pypi_0
nvidia-curand-cu12=10.3.2.106=pypi_0
nvidia-cusolver-cu12=11.4.5.107=pypi_0
nvidia-cusparse-cu12=12.1.0.106=pypi_0
nvidia-nccl-cu12=2.19.3=pypi_0
nvidia-nvjitlink-cu12=12.4.127=pypi_0
nvidia-nvtx-cu12=12.1.105=pypi_0
oauthlib=3.2.2=pypi_0
omegaconf=2.1.0=pypi_0
opencv-python=4.9.0.80=pypi_0
openssl=3.2.1=hd590300_1
pandas=1.4.4=pypi_0
pathtools=0.1.2=pypi_0
pip=24.0=pyhd8ed1ab_0
pkgutil-resolve-name=1.3.10=pypi_0
plotly=5.20.0=pypi_0
polars=0.20.18=pypi_0
promise=2.3=pypi_0
protobuf=3.19.0=pypi_0
psutil=5.9.8=pypi_0
pyarrow=15.0.2=pypi_0
pyasn1=0.6.0=pypi_0
pyasn1-modules=0.4.0=pypi_0
pydeprecate=0.3.1=pypi_0
pyntcloud=0.3.1=pypi_0
pyparsing=3.1.2=pypi_0
pyproj=3.5.0=pypi_0
python=3.8.19=hd12c33a_0_cpython
pytorch-lightning=1.8.2=pypi_0
ray=2.10.0=pypi_0
readline=8.2=h8228510_1
referencing=0.34.0=pypi_0
rich=12.4.4=pypi_0
rpds-py=0.18.0=pypi_0
safetensors=0.4.2=pypi_0
scikit-learn=1.3.2=pypi_0
scipy=1.9.3=pypi_0
sentry-sdk=1.44.1=pypi_0
setproctitle=1.3.3=pypi_0
setuptools=69.2.0=pypi_0
shapely=2.0.3=pypi_0
shortuuid=1.0.13=pypi_0
six=1.16.0=pypi_0
smmap=5.0.1=pypi_0
sympy=1.12=pypi_0
tenacity=8.2.3=pypi_0
tensorboard=2.11.2=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
termcolor=2.4.0=pypi_0
threadpoolctl=3.4.0=pypi_0
timm=0.9.16=pypi_0
tk=8.6.13=noxft_h4845f30_101
torch=1.11.0+cu113=pypi_0
torch-cluster=1.6.3=pypi_0
torch-geometric=2.2.0=pypi_0
torch-scatter=2.1.2=pypi_0
torch-sparse=0.6.16=pypi_0
torch-spline-conv=1.2.2=pypi_0
torchmetrics=0.10.2=pypi_0
torchvision=0.12.0+cu113=pypi_0
triton=2.2.0=pypi_0
typing-extensions=4.10.0=pypi_0
urllib3=1.26.18=pypi_0
wandb=0.13.5=pypi_0
wheel=0.43.0=pyhd8ed1ab_1
xz=5.2.6=h166bdaf_0
yarl=1.9.4=pypi_0
zipp=3.18.1=pypi_0
